// src/app/page/master/jenis-alamat/jenis-alamat.interface.ts
import {EventEmitter} from '@angular/core';

export interface JenisAlamat {
    kodeGolonganPegawai?: string;
	namaGolonganPegawai?: string;
	reportDisplay?: string;
	kodeExternal?: string;
	namaExternal?: string;
	noUrut?: string;
	statusEnabled?: boolean;
	id?: any;
	kdProfile?: string;
	kode?: any;
	version?: any;
	Version?: any;
	kodeJenisAlamat?: any;
	namaJenisAlamat?: any;

}